I confirmed that the sampling range selected caused the hyperbolic functions to blow up, giving inaccurate comparisons. That range has been reduced.
The code had unaccounted for memory leaks when using the =, the *=, and the /= operators. This issue has been addressed.

I added more print statements comparing the absolute and relative errors of the different functions. Numerical results agree well.

I added a function to generate data for a surface of the complex plane for sin(x+iy) that writes to sin_data.txt

Before moving to Quaternions, the * operator needs to use stack based iteration, instead of recursion, for improved performance.