Added comparison with boost's quaternion library. I was very disappointed to discover that boost has not implemented the log function, fractional/complex powers, or any of the inverse trigonometric functions. Boost has this to say about it:
"""
There are several reasons to this: on the one hand, the equivalent of analytic continuation for quaternions ("branch cuts") remains to be investigated thoroughly (by me, at any rate...), and we wish to avoid the nonsense introduced in the standard by exponentiations of complexes by complexes (which is well defined, but not in the standard...)
"""
I will need to find another resource to test the accuracy of my implementations.

For the functions that are available, numerical results are in good agreement. Multiplication timings for nion, however, is deplorable. The cost of generality I suppose (or ignorance). Timings for the transcendental functions are surprisingly more efficient (except for the tangent functions). This indicates I should avoid multiplication wherever possible by using alternate expressions for the transcendental equations. No more conveniently defining tan = sin/cos.