I have made all the functions both constexpr and inline to encourage the compiler to further improve execution time.
I have modified a few of the transcendental functions to improve their execution times. Noticeably, the sin and cosine functions are noticeably faster than the standard libraries.

I hardcoded the multiplication operation for degree 2 complex numbers. This removes an extra recursion depth in the multiplication algorithm that improves the time to evaluate higher order complex numbers. At some point it will be worthwhile to hardcode the quaternion and octonion products as well.

I discovered an inaccuracy in how I was measuring the performance of my hypercomplex library with the standard complex library and boost libraries. I was only comparing the differences in the norms of the results from the different functions I am testing. However, a more direct comparison is to evaluate the sum of the differences of each complex component. Adjusting this error metric still showed a good numerical agreement.

This error metric helped avoid a commit of a fallacious workaround for multiplication. I tried to avoid the recursion of multiplication by representing the product as p*q=exp(log(q)+log(p)) and reducing the expression to p*q=|p||q|exp(i theta + j phi); in otherwords, a product of the polar components.

This approach worked with traditional complex numbers, but it did not work with quaternions, unfortunately.I believe my assumption that exp(i)*exp(j) = exp(i+j) is false for the hypercomplex numbers. I wish to revisit this representation of the product again in the future, but for now I will keep the cayley-dickson definition of hypercomplex products.