I found a new library to reference by Douglas Wilhelm Harder associated with University of Waterloo. I have good numerical agreement with all his functions for the sedenions. This gives me greater confidence to say that boost has a serious bug with its octonion library. I will have to file a ticket to let the developers know.
I have included a Fetch_Content to download and compile Harder's library for comparison with my library. To enable, pass -DTEST_LARGE_DEGREE=ON to CMake during compilation.

I have also hard-coded the products for the quaternions and the octonions. If I intend to go to ridiculously large orders for hypercomplex numbers, I should hard-code the sedenion products as well. It's 256 terms, however, so I do not plan to do this anytime soon.

We can just forget about hard-coding Trigintaduonions. I'm a graduate student in chemistry; I don't have time for that.