Templated assigment and constructors of nions to allow conversions between nions with different arithmetic types and different max sizes.
Also changed the product operator to somewhat reduce the overhead of breaking the nions into pairs. It is still awfully slow in comparison to when hard coding the products by hand. I had hoped allocation on the stack would have fixed this, but it appears not. I guess at somepoint I'll actually have to manually implement the trigintaduonions if I want to use `ultracomplex` numbers.