Heavily modified the code to now allow for heap allocation without the specification of the second `N` template parameter.
In the process of making these changes, I discovered that I had a few issues with my move and copy constructors where they were not being called properly and unnecessary copies were being made. The code is now significantly faster than it had been.

I need to make a bunch of formatting changes to the testing functions, since I have places where I mix up the variables for the absolute and relative errors. This is not a huge priority for me, however.

I still don't have great agreements for some of my inverse trignometric functions. I believe the cause comes from a disagreement of the order of operations for defining those functions. I get conflicting results on what the `true` result should be depending on which library I use so it is not clear to me what result I should be getting. I could look into this more, but it is also not a priority right now, especially since the most important trigonometric functions do work as intended.